修改为正确的数据类型

+++ src/chap04_simple_neural_network/tf2.0-exercise.py
@@ -76,11 +76,11 @@ def softmax_ce(logits, label):
     return loss
 
 # 生成测试数据，形状为 [10, 5] 的正态随机数
-test_data = np.random.normal(size=[10, 5])
+test_data = np.random.normal(size=[10, 5]).astype(np.float32)
 # 进行softmax转换
 prob = tf.nn.softmax(test_data)
 # 生成标签，每个样本只有一个类别为 1
-label = np.zeros_like(test_data)
+label = np.zeros_like(test_data, dtype=np.float32)
 label[np.arange(10), np.random.randint(0, 5, size=10)] = 1.0
 # 比较自定义的损失值和tf自带结果，误差小于 0.0001 则认为相等
 ((tf.reduce_mean(tf.nn.softmax_cross_entropy_with_logits(label, test_data))
@@ -106,7 +106,7 @@ def sigmoid_ce(logits, labels):
     return loss
 
 # 测试逻辑
-test_data = np.random.normal(size=[10])
+test_data = np.random.normal(size=[10]).astype(np.float32)
 labels = np.random.randint(0, 2, size=[10]).astype(np.float32)
 
 # 对比 TensorFlow  原始结果和自定义函数结果
